backend/requirements.txt
# Kalshi Market Prediction App - Python Dependencies

# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Google Gemini
google-generativeai>=0.3.0

# Web framework (for API)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Optional: Additional ML libraries
# xgboost>=2.0.0  # Uncomment if using XGBoost instead of Random Forest
# tensorflow>=2.14.0  # Uncomment if using neural networks

# Development dependencies
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
